fix(ci): green Build check on Ubuntu, Windows, macOS
- SwapParams: restore missing expectedPeerSwapPubKey/ctrClaimTxId fields (SwapTypes.h)
- SwapDaemon.cpp: include Sia/SiaHtlcScript.h for SIA hash-lock path
- MSVC: explicit uint64_t casts for uint128_t→uint64_t in FixedPoint.cpp, WalletGreen.cpp, OrderbookTypes.cpp, PoolOrderOrchestrator.cpp
- check.yml macOS: install secp256k1 (system build path with USE_VENDORED_SECP256K1=OFF)

+++ src/Common/FixedPoint.cpp
@@ -47,7 +47,7 @@ FixedPoint64 FixedPoint64::mul(FixedPoint64 o) const {
   int128_t result = ((int128_t)a_hi * b_hi) << 64;
   result += (int128_t)a_hi * b_lo;
   result += (int128_t)a_lo * b_hi;
-  result += (int128_t)(((uint128_t)a_lo * b_lo) >> 64);
+  result += int128_t((uint64_t)(((uint128_t)a_lo * b_lo) >> 64));
 
   return FixedPoint64(result, Raw{});
 }
